+++ native~/Runtime/src/Cesium3DTilesetImpl.cpp
@@ -495,6 +495,8 @@ void Cesium3DTilesetImpl::LoadTileset(
   contentOptions.ktx2TranscodeTargets =
       CesiumGltf::Ktx2TranscodeTargets(supportedFormats, false);
 
+  contentOptions.applyTextureTransform = false;
+
   options.contentOptions = contentOptions;
 
   this->_lastUpdateResult = ViewUpdateResult();